# Comparison expression creation fix