## Description
Why.
This PR fixes the issue that user-defined operators can be a better match than the ones used to create Catch2 expression (defined in `Decomposer` and `ExprLhs`).

What.
- Overload with forwarding reference is added for all the comparison operators used in expressions.
- This overload doesn't compile for bit field non-const reference, so another overload that takes all arithmetic types by value is added. It replaces `bool` overload where it existed before.
- Operators are now defined as hidden friends to fix compilation on GCC.

## GitHub Issues
Closes #2121

## Note
I'm sorry for not fully reading the contribution guidelines. I just have a problem and propose a solution. Feel free to edit my code.