@AlCash07 2021-01-19
I compared compilation time with the current master by running these commands:
```
rm -rf debug-build
cmake -Bdebug-build -H. -DCMAKE_BUILD_TYPE=Debug -DCATCH_BUILD_EXAMPLES=ON -DCATCH_BUILD_EXTRA_TESTS=ON -DCATCH_DEVELOPMENT_BUILD=ON
cmake --build debug-build --target SelfTest
```
Compilation time for separate targets I got after one run:

Before the change:
Catch2: 33 sec
SelfTest: 28 sec

After the change:
Catch2: 32 sec
SelfTest: 30 sec

Regression in total is around 1.5%, but one run is obviously not statistically significant.
The conclusion I'd make is that there is no significant regression.