@AlCash07 2021-01-22
Compared compilation time again using this command
```
hyperfine --prepare 'rm -rf debug-build | cmake -Bdebug-build -H. -DCMAKE_BUILD_TYPE=Debug -DCATCH_BUILD_EXAMPLES=ON -DCATCH_BUILD_EXTRA_TESTS=ON -DCATCH_DEVELOPMENT_BUILD=ON' 'cmake --build debug-build --target SelfTest'
```
Before the change:
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `cmake --build debug-build --target SelfTest` | 63.719 ± 0.803 | 62.096 | 64.572 | 1.00 |

After the change:
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `cmake --build debug-build --target SelfTest` | 64.315 ± 0.694 | 62.726 | 64.832 | 1.00 |