@horenmar 2021-01-24
That's not good measurement: you explicitly include the time needed to build the static library proper, which is unaffected. However, I've had time to collect some measurements on my own:

This PR:
```
$ hyperfine -w 2 --prepare 'find ../tests/SelfTest/ -type f -name "*.cpp" -exec touch {} +' 'ninja -j 1'
Benchmark #1: ninja -j 1
  Time (mean ± σ):     19.532 s ±  0.463 s    [User: 18.527 s, System: 0.987 s]
  Range (min … max):   18.215 s … 19.693 s    10 runs
```

current `devel`
```
$ hyperfine -w 2 --prepare 'find ../tests/SelfTest/ -type f -name "*.cpp" -exec touch {} +' 'ninja -j 1'
Benchmark #1: ninja -j 1
  Time (mean ± σ):     18.914 s ±  0.720 s    [User: 17.922 s, System: 0.975 s]
  Range (min … max):   17.845 s … 19.694 s    10 runs
```

So that's difference of roughly 600ms for the median, which is about 3% slowdown over the current status. Not great, not terrible...

I will also try what happens if, instead of doing SFINAE for taking things by-value for arithmetic types, we return to the old functionality of only taking bools by value, which should require less work from the compiler to accomplish.