@AlCash07 2021-01-25
I just realized that the new test is added here, and it would be fair to compare compilation times without this test.