@AlCash07 2021-05-26
@horenmar I've learned about the "hidden friends" idiom, so I moved operators inside the class. I repeated your measurements, on my laptop they produce:

Current `devel`:
```
Benchmark #1: ninja -j 1
  Time (mean ± σ):     29.597 s ±  0.284 s    [User: 26.704 s, System: 2.309 s]
  Range (min … max):   29.071 s … 29.917 s    10 runs
```

This PR before the update:
```
Benchmark #1: ninja -j 1
  Time (mean ± σ):     30.549 s ±  0.560 s    [User: 27.571 s, System: 2.375 s]
  Range (min … max):   29.827 s … 31.406 s    10 runs
```

This PR now:
```
Benchmark #1: ninja -j 1
  Time (mean ± σ):     29.581 s ±  0.491 s    [User: 26.668 s, System: 2.313 s]
  Range (min … max):   29.129 s … 30.670 s    10 runs
```