@horenmar 2021-06-07
Thanks, the new compile time performance looks pretty good.

However, I think the change in inferred types is wrong for two reasons:
* it stops the assertions from being used with non-movable-non-copyable types. (I thought they were represented in the tests already, but turns out they are only tested in matchers, not plain assertions.)
* It causes needless runtime overhead in types that can be moved/copied.

As to the concern for dangling refs, the critical part of the code is this
```cpp
catchAssertionHandler.handleExpr( Catch::Decomposer() <= __VA_ARGS__ );
```

any temporary created as part of `__VA_ARGS__` should be alive until `handleExpr` returns, at which point Catch2 no longer keeps around the references to them. This means that it is fine to keep the inferred types as references.